+++ app.go
@@ -20,7 +20,7 @@ import (
 	"time"
 )
 
-var version = "0.7.3"
+var version = "0.7.4"
 
 const (
 	maxCachedQueries = 512